renamed eslint/files module to eslint/details
