+++ index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 module.exports = {
-    lintJavascript: require('./tasks/eslint/files'),
+    lintJavascript: require('./tasks/eslint/details'),
     lintReport: require('./tasks/eslint/report'),
     startKarmaServer: require('./tasks/start-karma-server'),
     startKarmaRunner: require('./tasks/start-karma-runner'),
