+++ tasks/eslint/details.js
